+++ Packages/com.vareiko.foundation/Runtime/Observability/ObservabilityConfig.cs
@@ -8,6 +8,7 @@ namespace Vareiko.Foundation.Observability
         [SerializeField] private FoundationLogLevel _minimumLogLevel = FoundationLogLevel.Info;
         [SerializeField] private bool _logToUnityConsole = true;
         [SerializeField] private bool _captureUnhandledExceptions = true;
+        [SerializeField] private bool _forwardUnhandledExceptionsToCrashReporting = true;
         [SerializeField] private bool _transitionToErrorStateOnUnhandledException = true;
         [SerializeField] private bool _enableDiagnosticsOverlay;
         [SerializeField] private float _diagnosticsRefreshIntervalSeconds = 0.25f;
@@ -15,6 +16,7 @@ namespace Vareiko.Foundation.Observability
         public FoundationLogLevel MinimumLogLevel => _minimumLogLevel;
         public bool LogToUnityConsole => _logToUnityConsole;
         public bool CaptureUnhandledExceptions => _captureUnhandledExceptions;
+        public bool ForwardUnhandledExceptionsToCrashReporting => _forwardUnhandledExceptionsToCrashReporting;
         public bool TransitionToErrorStateOnUnhandledException => _transitionToErrorStateOnUnhandledException;
         public bool EnableDiagnosticsOverlay => _enableDiagnosticsOverlay;
         public float DiagnosticsRefreshIntervalSeconds => _diagnosticsRefreshIntervalSeconds <= 0f ? 0.25f : _diagnosticsRefreshIntervalSeconds;